# backend/tests/snapshots/openapi_schema.json
{
  "components": {
    "schemas": {
...
      "CheckoutResponse": {
        "description": "Response for POST /checkout.",
        "properties": {
          "checkout_url": {
            "title": "Checkout Url",
            "type": "string"
          }
        },
        "required": [
          "checkout_url"
        ],
        "title": "CheckoutResponse",
        "type": "object"
      },
...
      "DimensionItem": {
        "properties": {
          "count": {
            "title": "Count",
            "type": "integer"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "value": {
            "title": "Value",
            "type": "number"
          }
        },
        "required": [
          "name",
          "count",
          "value"
        ],
        "title": "DimensionItem",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "SearchesOverTimeResponse": {
        "properties": {
          "data": {
            "items": {
              "$ref": "#/components/schemas/TimeSeriesDataPoint"
            },
            "title": "Data",
            "type": "array"
          },
          "period": {
            "title": "Period",
            "type": "string"
          }
        },
        "required": [
          "period",
          "data"
        ],
        "title": "SearchesOverTimeResponse",
        "type": "object"
      },
...
      "SummaryResponse": {
        "properties": {
          "avg_results_per_search": {
            "title": "Avg Results Per Search",
            "type": "number"
          },
          "estimated_hours_saved": {
            "title": "Estimated Hours Saved",
            "type": "number"
          },
          "member_since": {
            "title": "Member Since",
            "type": "string"
          },
          "success_rate": {
            "title": "Success Rate",
            "type": "number"
          },
          "total_downloads": {
            "title": "Total Downloads",
            "type": "integer"
          },
          "total_opportunities": {
            "title": "Total Opportunities",
            "type": "integer"
          },
          "total_searches": {
            "title": "Total Searches",
            "type": "integer"
          },
          "total_value_discovered": {
            "title": "Total Value Discovered",
            "type": "number"
          }
        },
        "required": [
          "total_searches",
          "total_downloads",
          "total_opportunities",
          "total_value_discovered",
          "estimated_hours_saved",
          "avg_results_per_search",
          "success_rate",
          "member_since"
        ],
        "title": "SummaryResponse",
        "type": "object"
      },
      "TimeSeriesDataPoint": {
        "properties": {
          "label": {
            "title": "Label",
            "type": "string"
          },
          "opportunities": {
            "title": "Opportunities",
            "type": "integer"
          },
          "searches": {
            "title": "Searches",
            "type": "integer"
          },
          "value": {
            "title": "Value",
            "type": "number"
          }
        },
        "required": [
          "label",
          "searches",
          "opportunities",
          "value"
        ],
        "title": "TimeSeriesDataPoint",
        "type": "object"
      },
      "TopDimensionsResponse": {
        "properties": {
          "top_sectors": {
            "items": {
              "$ref": "#/components/schemas/DimensionItem"
            },
            "title": "Top Sectors",
            "type": "array"
          },
          "top_ufs": {
            "items": {
              "$ref": "#/components/schemas/DimensionItem"
            },
            "title": "Top Ufs",
            "type": "array"
          }
        },
        "required": [
          "top_ufs",
          "top_sectors"
        ],
        "title": "TopDimensionsResponse",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "description": "API para busca e an\u00e1lise de licita\u00e7\u00f5es de uniformes no PNCP. Permite filtrar oportunidades por estado, valor e keywords, gerando relat\u00f3rios Excel e resumos executivos via IA.",
    "title": "BidIQ Uniformes API",
    "version": "0.2.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/analytics/searches-over-time": {
      "get": {
        "description": "Get time-series search data grouped by period.",
        "operationId": "get_searches_over_time_analytics_searches_over_time_get",
        "parameters": [
          {
            "in": "query",
            "name": "period",
            "required": false,
            "schema": {
              "default": "week",
              "pattern": "^(day|week|month)$",
              "title": "Period",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "range_days",
            "required": false,
            "schema": {
              "default": 90,
              "maximum": 365,
              "minimum": 1,
              "title": "Range Days",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SearchesOverTimeResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get Searches Over Time",
        "tags": [
          "analytics"
        ]
      }
    },
    "/analytics/summary": {
      "get": {
        "description": "Get overall user analytics summary.",
        "operationId": "get_analytics_summary_analytics_summary_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SummaryResponse"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get Analytics Summary",
        "tags": [
          "analytics"
        ]
      }
    },
    "/analytics/top-dimensions": {
      "get": {
        "description": "Get top UFs and sectors by search count.",
        "operationId": "get_top_dimensions_analytics_top_dimensions_get",
        "parameters": [
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 5,
              "maximum": 50,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TopDimensionsResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get Top Dimensions",
        "tags": [
          "analytics"
        ]
      }
    },
...
    "/checkout": {
      "post": {
        "description": "Create Stripe Checkout session for a plan purchase.",
        "operationId": "create_checkout_checkout_post",
        "parameters": [
          {
            "in": "query",
            "name": "plan_id",
            "required": true,
            "schema": {
              "title": "Plan Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "billing_period",
            "required": false,
            "schema": {
              "default": "monthly",
              "title": "Billing Period",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CheckoutResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Create Checkout",
        "tags": [
          "billing"
        ]
      }
    },
...
    "/v1/analytics/searches-over-time": {
      "get": {
        "description": "Get time-series search data grouped by period.",
        "operationId": "get_searches_over_time_v1_analytics_searches_over_time_get",
        "parameters": [
          {
            "in": "query",
            "name": "period",
            "required": false,
            "schema": {
              "default": "week",
              "pattern": "^(day|week|month)$",
              "title": "Period",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "range_days",
            "required": false,
            "schema": {
              "default": 90,
              "maximum": 365,
              "minimum": 1,
              "title": "Range Days",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SearchesOverTimeResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get Searches Over Time",
        "tags": [
          "analytics"
        ]
      }
    },
    "/v1/analytics/summary": {
      "get": {
        "description": "Get overall user analytics summary.",
        "operationId": "get_analytics_summary_v1_analytics_summary_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SummaryResponse"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get Analytics Summary",
        "tags": [
          "analytics"
        ]
      }
    },
    "/v1/analytics/top-dimensions": {
      "get": {
        "description": "Get top UFs and sectors by search count.",
        "operationId": "get_top_dimensions_v1_analytics_top_dimensions_get",
        "parameters": [
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 5,
              "maximum": 50,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TopDimensionsResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get Top Dimensions",
        "tags": [
          "analytics"
        ]
      }
    },
...
    "/v1/checkout": {
      "post": {
        "description": "Create Stripe Checkout session for a plan purchase.",
        "operationId": "create_checkout_v1_checkout_post",
        "parameters": [
          {
            "in": "query",
            "name": "plan_id",
            "required": true,
            "schema": {
              "title": "Plan Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "billing_period",
            "required": false,
            "schema": {
              "default": "monthly",
              "title": "Billing Period",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CheckoutResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Create Checkout",
        "tags": [
          "billing"
        ]
      }
    },
...
  }
}